Fix: Move bean creating from main class to configuration class
