+++ stats/stats-service/src/main/java/ru/practicum/ewm/StatsService.java
@@ -2,19 +2,11 @@ package ru.practicum.ewm;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-
-import java.time.Clock;
 
 @SpringBootApplication
 public class StatsService {
 
     public static void main(String[] args) {
         SpringApplication.run(StatsService.class, args);
-    }
-
-    @Bean
-    Clock clock() {
-        return Clock.systemDefaultZone();
     }
 }
