+++ stats/stats-service/src/test/java/ru/practicum/ewm/configuration/ClockConfig.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.common;
+package ru.practicum.ewm.configuration;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
